Adding more swagger auth tests
